refactor: 금액 범위 int -> double 변경
할인율 계산 정확도를 위해 int -> double 변경

+++ src/main/java/store/domain/DiscountStrategy.java
@@ -1,5 +1,5 @@
 package store.domain;
 
 public interface DiscountStrategy {
-    int getDiscountPrice(int price);
+    double getDiscountPrice(double price);
 }
